#211 - lib: vector cache framework
