+++ src/libenv/core/reader.rs
@@ -15,7 +15,7 @@ use crate::{
         stream::{Core as _, Stream},
         struct_::{Core as _, Struct},
         symbol::{Core as _, Symbol},
-        vector::{Core as _, Vector},
+        vectors::{Core as _, Vector},
     },
 };
