+++ src/libenv/streams/read.rs
@@ -18,7 +18,7 @@ use crate::{
         cons::{Cons, Core as _},
         stream::{Core as _, Stream},
         symbol::{Core as _, Symbol},
-        vector::{Core as _, Vector},
+        vectors::{Core as _, Vector},
     },
 };
