+++ src/libenv/types/stream.rs
@@ -22,8 +22,8 @@ use crate::{
     types::{
         char::Char,
         symbol::{Core as _, Symbol},
-        vecimage::{TypedVec, VecType},
-        vector::Core as _,
+        vector::{TypedVec, VecType},
+        vectors::Core as _,
     },
 };
